Format generated code with rustfmt/prettier
This uncovered a bunch of issues where syntactically valid code is
generated. Most of the cases are fixed by ensuring that `name != ""`.
I think that the name of the routebuilder should be passed in as a
mandantory constructor argument to `new` to avoid this.

Remaining edge cases will be fixed in following commits.